pyproject.toml
[project]
name = "streaming-data-loader"
version = "0.1.0"
description = "Tool for efficient loading and processing of streaming data from Kafka to Elasticsearch"
authors = [
    { name = "aDudko", email = "anatoly_dudko@icloud.com" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-json-logger (>=3.3.0,<4.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "aiokafka (>=0.12.0,<0.13.0)",
    "elasticsearch (>=8.17.2,<9.0.0)",
    "aiohttp (>=3.11.14,<4.0.0)",
    "pip (>=25.0.1,<26.0.0)",
    "pytest-asyncio (>=0.25.3,<0.26.0)",
    "prometheus-client (>=0.21.1,<0.22.0)"
]

[tool.poetry]
packages = [{ include = "streaming-data-loader", from = "src" }]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
